[Keys] Make type=cycle automatically synchronise to current
Our type=cycle and type=toggle key bindings can get out of sync in many
non-trivial situations, such as if multiple keys, presets and/or command
lists can set the same variable.

This adds some smarts to the type=cycle key bindings to check if the
current reality (or transition target) matches their current preset, and
if it doesn't they check to see if any of the other presets do match,
and if one is found they will jump to that preset before cycling. If no
presets match it will continue as before starting with the first preset.

This improved behaviour is desirable in most cases, so this is set as
the default. smart=false can be specified to return to the old
behaviour.

The same is not done for type=toggle, because the transition
save/restore complicates things and the toggle off value is not
specified (we might assume 0, but that's not always right), but a two
phase type=cycle can easily be used in its place.