Smart cycles: Deal with nvapi separation/convergence munging
NvAPI doesn't return the exact same separation / convergence value we
set, so checking for an exact match will never work (even values that
can be exactly represented in floating point can still be munged). This
adds a tolerance to both separation and convergence checks to match
whenever it is close enough, and adds an extra special case to deal with
the fact that nvapi caps separation at 1% under normal circumstances.