Command List: Make "pre/post run=" run pre & post lists together
The run= command usually runs the target pre command list when the run
command itself is run in a pre command list, and runs the target post
command list when it itself is run in a post command list. For the most
part this is what we want so that pre and post maintain their meaning
down the chain with respect to the original draw call, but there are a
few exceptions where it makes more sense to run the target pre and post
lists together.

We already do this for run=CustomShader, as the meaning of pre and post
in those sections is logically (if not technically) more relative to the
injected draw call. However when we do "pre run = CommandList" or "post
run = CommandList" only one phase of the target command list will
actually get run, which tends to just be more confusing than actually
useful. If pre or post was specified in the run command than logically
it makes more sense that pre and post should be combined together in the
target command list so they run one after the other.

This makes the [Constants] post command list a little simpler when
calling a command list to perform some processing after the
d3dx_user.ini file has been loaded, so that the called command lists
don't have to put the post keyword everywhere to work.