Keys: Remove character limits from type=cycle options
Code that has not yet moved to the newer ini parsing APIs with C++
strings still have the old 260 character limit. For most things that is
good enough, but type=cycle key bindings can potentially go beyond that,
especially if run= is being specified with a command list per cycle.

This switches this code to use the newer APIs, removing the character
limit. Since the newer APIs intentionally do not return wide character
strings it also switches a lot of the supporting code to use regular
characters instead of wide characters, and adds a few missing utility
routines for this.