Keys/type=cycle: Delay post command lists until after UpdateTransitions
This delays the post phase of type=activate/cycle key bindings until
after the UpdateTransitions() routine has run, allowing these command
lists to see the updated values, which might be handy for e.g. firing
off a notification shader after updating a type=cycle value based on the
newly set value.

If the values are transitioning they will see the value in the first
frame of the transition, so this is probably more useful for instant
values.