Add key binding to wipe the d3dx_user.ini file
There are often times when we might want to quickly check how a fix
behaves when loaded with no persistent configuration, or perhaps to wipe
out a persistent config file that is throwing errors (e.g. maybe a
variable that used to be persistent was removed in an update or the ini
file it belongs to is no longer included).

However this is not always as simple as deleting the d3dx_user.ini file
and reloading, because something might be continually updating a
persistent variable causing 3DMigoto to write it out again on the next
config reload.

This commit adds a new key binding that will wipe out the d3dx_user.ini
file and reload the config for these cases.