Input: Only process input if the game window is foregrounded
This adds an option to check if the game is foregrounded before
processing any keyboard / controller input to avoid inadvertently
changing settings while alt+tabbed out of the game.

The setting is currently disabled by default if not specified in the
d3dx.ini until we know if there are any potential issues with games that
do weird stuff with their windows. The template d3dx.ini does have it
enabled by default so that new fixes will pick it up.