Profiling: Add statistics of a wide range of GPU impacting tasks
This adds statistics of various tasks that may have an impact on the GPU
performance that won't (necessarily) show up in the existing CPU
performance measurements.

In particular it adds counts of all the various types of resource copies
that are performed each frame, as well as statistics about the resource
pools, injected / skipped draw calls and IniParams updates.

Also add CPU timing measurements of the software cursor management since
that has shown up as a hotspot in profiling elsewhere, even if it
reportedly has not impacted the framerate in those cases.

Also add ShaderRegex timing measurements as that is a strong candidate
for moving out of the draw call since the deferred requirements are for
as yet unmerged code (constant buffer / other render state matching).