Profiling: Warn if post checktextureoverrides were not optimised out
checktextureoverride can add overhead in both pre and post command
lists, but in a lot of cases the post command list is unnecessary.
Because of the way this works, 3DMigoto can only optimise out the post
command lists if none of the TextureOverride sections contain post
commands, but because some commands are implicitly added to both command
lists it is not always obvious when this has happened.

This adds a notice to the profiling overlay to indicate when this has
happened, and indicate which commands were responsible.