Command List: Add operands for hunting and frame analysis
The hunting operand can be used to check if hunting is enabled and
whether the overlay is displayed. This is mostly intended for frame
analysis profile key bindings (with condition=hunting) so that they will
be inactive in release mode.

The frame_analysis operand can be used to check if frame analysis is
currently dumping. This is intended to allow mesh swaps to be
temporarily disabled during frame analysis to dump the original meshes
instead of the modified ones.