Fix crash in [key] condition expression parsing
find_local_variable() would unconditionally dereference the scope
pointer, causing a NULL pointer dereference if the expression was
outside of the scope of a command list, e.g. the condition expression in
[Key] bindings.

Local variables cannot exist outside of a command list, so always bail
in this case. Also add an early "doesn't look like a variable" exit to
this routine to skip looking for local variables when the syntax is
obviously not one.