Command list: Make "pre/post checktextureoverride" run pre & post together
This follows the same idea as the change to the explicit run command,
such that if checktextureoverride is run with an explicit pre or post it
will run both the pre and post lists of the target [TextureOverride]
lists.

Note that this will not stop any of these lists from preventing the
implicit post checktextureoverrides from being optimised out since there
is no way to determine at parse time which lists may be called from
which checktextureoverrides.