checktextureoverride: Make "this" refer to the resource slot being checked
This allows [TextureOverride] command lists to refer to the resource
they were checked with regardless of which slot it is in. This means we
no longer have to rely on [TextureOverride]'s having to know which slot
they will be used with a checktextureoverride command, and allows them
to work when the texture may appear in various different slots.