Deduce bind_flags at runtime when copying by reference
bind_flags is usually deduced at parse time, but there are some
circumstances where this cannot happen: If one custom resource is copied
by reference to another, and where the "this" target is assigned in a
[TextureOverride].

Until now these cases have required bind_flags to be manually specified
in the custom resource, but that is problematic as it requires the
modder to know more about DirectX than I would like them to need to
know, and in some cases using the wrong bind flags seems to work
regardless, possibly masking bugs that may appear on other systems.

This tries to solve the most common cases of this by using the bind
flags of a reference copy target whenever a custom resource is
substantiated, thereby deducing the required bind flags when the
resource is first used if they could not be deduced earlier.

There are still some complex cases where this may fall short - for
instance if a custom resource was copied by reference to another, and
that second custom resource was sometimes bound as a shader resource,
and at other times bound as a render target or UAV. It could also happen
if a [TextureOverride] replacing "this" can be checked as both a texture
and render target/UAV at different times. But these remaining cases
should be sufficiently rare that most people will never have to worry
about it, and if they do crop up the bind flags can still be manually
overridden.