Fix [Key] condition case sensitivity
Expressions are supposed to be case insensitive, which is true for any
expressions parsed in the command list since the entire command list is
case insensitive, but the expression the [Key] condition option was
missing this. Transform these expressions to lower case before parsing
them to make them case insensitive.