Frame Analysis Log: Fix newline in wrong position
Several of the Clear***View functions had the position of the newline
and close bracket mixed up, causing some minor issues in the frame
analysis and debug logs.

This flips them the right way around, and also adds some indentation to
the detailed view info printed out in these and similar functions.