Custom Resources: Transfer resources between devices
DOAXVV creates new temporary DirectX devices & contexts whenever the
news is displayed, which results in 3DMigoto incorrectly re-running the
[Constants] command list (fixing that is left for later). If a mod uses
this command list to copy resources they will be copied using the wrong
DirectX device, resulting in a crash.

The previous two commits tackled part of this issue, by discarding
resources on the source and destination sides of the copy with a device
mismatch to fix the crash, however there were cases where this would
result in a resource losing it's contents. In practice this resulted in
certain mods appearing to stop working (original meshes from game being
rendered since the replacement vertex & index buffers were now NULL)
until the next config reload re-initialised those buffers.

This patch adds a procedure to perform an inter-device resource copy to
transfer the resources from the old device to the new. It does this by
copying the resource to a temporary staging resource, mapping each
contained subresource to the CPU, then using that to update a
corresponding subresource on a new resource created on the destination
device.

This resolves the remaining critical issues in DOAXVV - the affected
no longer crash, and with the affected index & vertex buffers retain
their contents after viewing the news.