Do not re-run [Constants] on new device/context creation
DOAXVV creates new temporary DirectX devices & contexts whenever the
news is displayed, which causes the [Constants] command list to be
re-run, and can reset variables and resources used by the mods back to
their default / most recently saved values.

[Constants] is only intended to be run once on launch and again after
any config reload, so add a check to limit it to only run at those
times.

I could conceive of some situations where we might prefer to go the
opposite way and reinitialise everything when a new device is created,
and this would simplify cases where a resource created on one device is
later used on another (since we could discard all custom resources,
caches, etc and just start over), but this would mostly be limited to
cases where the device switching happens at launch when it won't matter
if variables are lost, and therefore is not a good generic solution.