Check for COM identity violations in non Hacker Objects
We currently rely on the COM identity rule to locate our HackerDevice
if any of the CreateSwapChain methods are called with an unwrapped
device, however when using ReShade in Resident Evil 2 this fails because
ReShade violates the COM identity rule (as do we ;-).

Our "throw our hands in the air and shrug I dunno" methodology involves
outputting a list of all known interfaces that the mystery device does
support to the log file for further analysis. Since COM identity
violations are now in the list of things that can cause this, check for
those as well.

In RE2 with ReShade 4.1.1 this outputs:

    HackerDevice 000000005BF457A0 created to wrap 000000005BEBF140
    ...
    lookup_hacker_device(000000005CCE7700) IUnknown: 000000005CCE7700 HackerDevice: 0000000000000000
    WARNING: Could not locate HackerDevice for 000000005CCE7700
    Checking what interfaces 000000005CCE7700 supports...
      Supports IUnknown: 000000005CCE7700
      Supports IDXGIDevice: 000000005CCE7700
      Supports IDXGIDevice1: 000000005CCE7700
      Supports IDXGIDevice2: 000000005CCE7700
      Supports IDXGIObject: 000000005CCE7700
      Supports ID3D10Multithread: 000000005BF4F810 (COM identity violation: 000000005BF4EFB0)
      Supports ID3D11Device: 000000005BEBF140 (COM identity violation: 000000005BEBF140)
      Supports ID3D11Device1: 000000005BEBF140 (COM identity violation: 000000005BEBF140)
      Supports IDXGIDevice3: 000000005CCE7700
      Supports ID3D11Device2: 000000005BEBF140 (COM identity violation: 000000005BEBF140)
      Supports IDXGIDevice4: 000000005CCE7700
      Supports ID3D11Device3: 000000005BEBF140 (COM identity violation: 000000005BEBF140)
      Supports ID3D11Device4: 000000005BF4F7E0 (COM identity violation: 000000005BF4EFB0)
      Supports ID3D11Multithread: 000000005BF4F810 (COM identity violation: 000000005BF4EFB0)
      Supports ID3D11Device5: 000000005BF4F7E0 (COM identity violation: 000000005BF4EFB0)
    BUG: Unwrapped ID3D11Device!

In this case the DXGI interfaces and the IUnknown queried from those
interfaces resolve to a ReShade object, however the D3D11 devices and
IUnknown queried from those interfaces resolve to the DirectX object,
which is in violation of the COM identity rule as all queries for
IUnknown must resolve to the same physical pointer regardless of which
interface was used for the query.

That said, this violation is somewhat expected for a 3rd party tool
wrapping DirectX (and 3DMigoto is also guilty of violating the same
rule), so we will need to handle this situation.

Somewhat more alarming is that this has determined that the Device4, 5
and Multithread interfaces also violate the COM identity rule, and I
doubt that they would have anything to do with ReShade, meaning it
appears that Microsoft has violated their own rule.