Implement method to find HackerDevice impervious to 3rd party tools
In some games (e.g. Resident Evil 2) when using ReShade, CreateSwapChain
is passed an IUnknown device that we do not recognise (presumably a
device wrapped by ReShade), that doesn't respond to our QueryInterface
for HackerDevice (so, not 3DMigoto device wrapped in a ReShade device,
or the QueryInterface has been blocked), and who's IUnknown does not map
to the DirectX identity (violating the COM identity rule).

This implements a hopefully full proof backup strategy to locate our
HackerDevice, by storing a pointer to it in the DirectX device's private
data when it is first created that we can look up from the
CreateSwapChain call.