Fix offsetof compiler warning in vs2017 toolchain
When compiled with the latest vs2017, the following warning is thrown:

    warning C4644: usage of the macro-based offsetof pattern in constant
    expressions is non-standard; use offsetof defined in the C++
    standard library instead (compiling source file Override.cpp)

Switch the code to use the offsetof macro as suggested to fix the
warning. Note that the line that converts the pointer to member variable
into an offset still uses the existing syntax (since the macro can't
work with pointer to members), but that line does not trigger the
warning since it is not a constant expression.