Avoid calling nvapi with a NULL stereo handle
Running with a debugger attached showed that calling into nvapi with a
NULL stereo handle (e.g. when 3D Vision is disabled in the control
panel) would result in a NULL pointer dereference exception being
thrown.

nvapi must catch this exception since we don't ordinarily see a crash,
but even if that is the case exceptions aren't great for performance and
makes debugging sessions annoying, so just avoid calling nvapi with a
NULL stereo handle and directly return NVAPI_ERROR.