TestShaders: Add failing cases from Splinter Cell Blacklist
export_hlsl leads to a crash (not a catchable exception) in this game,
preliminarily narrowed down to an issue with Hull shaders.
cmd_decompiler doesn't crash, but does throw exceptions on a number of
hull shaders, so add one of these as an example for our decompiler test
suite.

Also add a random pixel shader with an unknown opcode.