HLSL Test Suite: Add support for using existing .bin files
For any new game examples we really want to use the .bin files rather
than attempting to reconstruct them to make sure we are testing a shader
from the wild and not a possibly botched lab reconstruction.

This allows the game example test script to be run on a .bin file
skipping the reconstruction step.

This adds the Hellblade test shader that was added earlier, but does not
yet add the Splinter Cell Blacklist shaders since they fail to
decompile.