cmd_Decompiler: Fail assembler verification for SHDR / SHEX mismatch
There are two possible signatures for the bytecode section in a shader
binary - SHDR and SHEX. SHEX is always used for shader model 5, and we
have made an assumption that SHDR would always be used for shader model
4. This turns out not to be the case as using certain features will
cause a model 4 shader to use SHEX as well (Presumably SHEX is short for
"SHader EXtension").

This mismatch led to a bug in cmd_Decompiler's assembler verification,
as it does not fail shaders for missing or extra sections (since RDEF,
STAT, etc are expected to be missing), but the mismatch means it won't
actually compare their contents.

In practice this means that a number of shader model 4 test cases pass
when they should in fact fail (specifically the debug & minimum
precision test cases) since the bytecode does not get verified. A number
of other test cases with mismatched sections also pass, though in their
case it is not clear whether the mismatch is serious or not (since the
SHDR and SHEX sections are identical in every way except their
signature), but since there is the potential that it could cause an
issue flag these as failures.