cmd_Decompiler: Fail shader verification if any non-whitelisted section is missing
To catch any cases where we might be missing some other important
section, fail shader verification if any non-whitelisted section is
missing. Currently only STAT and RDEF are whitelisted as known optional
sections.

This has flagged the input & output signatures on the same shaders that
had the SHDR/SHEX mismatch, because they will have used the older
signature sections for the same reason they used the SHDR section. If
this does turn out to be an issue we will need to upgrade these sections
(however, IIRC unlike SHDR vs SHEX there are some differences between
them so we can't simply swap the signature after the fact).

This has also flagged a section called "SFI0" (with a zero) present in
certain shaders that I have not yet checked to determine what it is or
whether it is optional.