Assembler: Break assembleOp up into several functions
Split out the cases to assemble literals and indexable registers into
separate functions to make assembleOp a little easier to follow.

This also adds a union to the token_operand structure so that the
operand can be accessed through it rather than having to use a separate
type cast pointer.