Assembler: Add support for minimum precision tags
In Resident Evil 2 we see a number of shaders using minimum precision
types, which sets an extended flag (that we already handle) and
introduces optional tags after each operand such as the following:

    dcl_input_ps linear v1.xyzw {min16f}
    dp2 r2.x {min16f}, v2.xyxx {min16f}, v2.xyxx {min16f}
    movc r0.xyzw {min16f}, r0.xxxx, l(0,0,0,0) {def32 as min16f}, v1.xyzw {min16f}
    mov o0.xyzw, r0.xyzw {min16f as def32}

These essentially boil down to two variations on the tag: {minXX} is the
simple case indicating that the operand is a minimum precision type and
{typeXX as typeYY} indicating that the operand of typeXX needs to be
cast to match the destination of typeYY.

This commit modifies the assembler to group this tag with the operand it
corresponds with, and to parse the tag and encode the type into the
appropriate bits in the operand. It turns out that we can pretty much
ignore the "... as type2" part of the syntax in source operands since
that is implied by the encoding in the destination operand (verified by
extending the test case to include this case and confirmed on RE2
shaders).

The minimum precision test cases now pass their SHDR/SHEX comparison,
though they are currently still listed as failing overall due to the
missing SFI0 section, the shader model 4 variant is using SHDR
instead of SHEX and both are using the wrong output signature version.
It is not yet clear if any of these matter in practice.