cmd_Decompiler: List sections in new shader not found in old
When validating the dis[assembler], if we have a section version
mismatch we print out the section that was missing, but don't indicate
which section was present, which makes it not immediately obvious what
the issue is. This prints a list of all sections found in the new shader
that were not in the old to make it more obvious.