cmd_Decompiler & asm test suite: Add --lenient option
This adds an option to relax the conditions that will fail the assembler
verification back to what it was before, allowing us to check if the
assembler is working without worrying about section version mismatches.

This adds the same option to the test suite so that we can determine if
a test is failing due to a bug in the assembler, or due to section
version mismatches.