Assembler: Handle absolute value and minimum precision combined
When minimum precision and absolute value are combined, the operand
looks like this:

    // mad r2.xyzw {min16f}, |r2.xyzw {min16f}|, l(-4.00000...

This fails since the operand will be cut short at the closing brace, and
again because the minimum precision parser will truncate the operand at
the opening brace.

This patch changes the minimum precision handling to split at the comma
so that the closing absolute value bar is included in the operand, and
removes the minimum precision tag once it has been processed, leaving
the closing bar intact.

It also processes the minimum precision tag after the absolute value
bars as a secondary strategy of achieving the same thing.

The following test cases now pass with --lenient:
    GameExamples/re2/03cdbb1d64be7c53-cs.bin
    GameExamples/re2/1d62a8c00ed1f398-cs.bin