Assembler: Add initial support for double precision literals
These are used in some shaders in Resident Evil 2. They are similar to
the l(val1, val2, val3, val4) construct for floating point literals, but
in the form f(val1, val2).

This adds the necessary parsing to handle simple cases of these, though
NAN literals and the like are not handled, and there is no precision
fixup path for these as yet.

With this change, all shaders in Resident Evil 2 now successfully
compile, though the min & double precision shaders still require
--lenient to pass the test case.