Assembler: Add the first parse error exception
This adds an exception the assembler can throw if it encounters any
parse error and hooks it up to the double literal handling if the
expected comma between the two values is missing.

This is intended to be a place to start adding exception handling
throughout the rest of the assembler.