Assembler: Extend precision fixup to include doubles
Like floats, the MS disassembler uses %f for doubles, which can lead to
a loss of precision for small numbers, or numbers with a large number of
significant digits.

This extends the assemble and compare strategy used to correct these
literals to cover doubles as well. It takes a simpler approach, just
using %g to let libc switch between readable and scientific notation as
required, with DBL_DECIMAL_DIG level of precision, which is guaranteed
to allow us to recreate the original value exactly.

This also handled inf & nan cases, by replacing them with hex codes that
can then be parsed. Later we can add parsing for the actual inf/nan
literals that the disassembler generates if desired.