Test suite: Run assembler test on HLSL compiled GameExample tests
This runs the assembler tests on shaders compiled from HLSL source
during shader reconstruction. Since these have not themselves gone
through the assembler they make better test cases for the assembler than
the reconstructed shaders that we use to test the decompiler.

This has located a single failure in one of our example shaders from
WatchDogs (and indeed this failure does not show up in the reconstructed
version).