Disassembler precision fixup: Minor refactoring
This replaces the giant all encompassing if block inside the loop with
an inverse test and continue to save a level of indentation.

Also noted that the precision fixup code seems to have a case to bail if
debug info is present, but we should probably try to cope with that
since there are games in the wild that have debug info present and we
still want the precision fixup to run on those.