cmd_Decompiler: Add --disassemble-hexdump
This adds a new option to cmd_Decompiler to disassemble a shader
intermixing it with extra comments containing the raw hex values of each
instruction to make it easier to understand what has gone wrong in a
shader that is failing validation. This is vastly superior to the /Fx
option of fxc, as that does nothing more than appending a hexdump to the
end of the file.

In the event that shader validation fails, this hexdump will be run and
logged to stderr.

Internally this works by extending the disassembler wrapper we use to
fix precision issues since that already simultaneously parses the
assembly text and the binary making it an ideal spot to add intermixed
hexdump outputs.

Offsets are printed relative to the file to make it easier to locate
each instruction in a hex editor, which could be useful if we wanted to
manually hex edit a shader for any reason.