cmd_Decompiler: Only hexdump bad instructions on validation failure
Rather than hexdumping every instruction when a shader fails validation,
only hexdump specific instructions that fail validation.

This alters the routine a little to remove the distinction between the
's' (instruction) and 'sNew' (precision fixed instruction) to allow the
hexdumper to use the precision fixed version so it doesn't highlight
MS's bugs, and refactors the hexdump into a separate function.