Disassembler: Fix precision fixup & hexdump when debug info is present
When debug info is present in a shader we currently bail out of the
disassembler's precision fixup path, which means we may end up with
precision issues left in the disassembled shader, and cmd_Decompiler's
hexdump feature will fail and the shader will fail validation.

This passes the flag to disable the debug info when disassembling the
shader, thereby giving us the predictable assembly that we require to
match up with the bytecode for the precision fixup and hexdump.

It also whitelists the debug info SDBG section as okay to be omitted
when validating the shaders.

This allows the MGSV test case shader to pass validation.