cmd_Decompiler: Allow missing level 9 bytecode & add test cases
All these shaders from MGSV were still failing assembler validation, and
upon investigation it was determined that they include feature level 9
bytecode in a extra Aon9 section that we do not assemble. I believe this
is not critical, as these should be using the DX11 bytecode in the
SHDR/SHEX section, so allow this section to be omitted and still pass
the verification.

The only cases I've seen where a vs_3_0 shader was legitimately used in
DX11 was the Steam overlay (enable hooking to intercept), so it is
conceivable that we may need to deal with these some day, but unlikely.