Assembler: Drop precision fixup debugging
The assembler contains an obscure debugging feature that will dump out
any floating point values with precision issues to a random debug file
that no one knows about or ever checks. The validatation and hexdump
modes of cmd_Decompiler can serve the same purpose as this feature, so
drop it.