cmd_Decompiler: Don't add StereoParams & IniParams to decompiled shaders
cmd_Decompiler inserts StereoParams & IniParams declarations in the
usual t125 and t120 slots when decompiling a shader. This seemed like a
good idea at the time, but given that cmd_Decompiler is standalone from
3DMigoto it actually doesn't make a whole lot of sense that it would add
3DMigoto registers to decompiled shaders. In addition, this causes some
of our test cases to fail if they already have StereoParams defined
since it will now be defined twice.

This change may break some scripts that depend on the existing
behaviour, but those scripts can be updated or continue using an old
version.

This commit also does:

    run_game_example_tests.sh --hlsl --update-chk
    run_hlsl_tests.sh --update-chk

The following previously failing test cases are now passing:

    GameExamples/Cars/7cbe14ee3f1cddb1-vs_replace.txt
    GameExamples/Cars/ab6c5888854b6e44-vs_replace.txt
    GameExamples/Cars/tweak 7cbe14ee3f1cddb1-vs_replace.txt
    GameExamples/FC4/5845f1c6886ed626-vs_replace.txt
    GameExamples/MadMax/692960842b51655f-vs_replace.txt
    GameExamples/MadMax/d09b2165321f05cf-ps_replace.txt
    GameExamples/SR3/72bd9b1f9ed176ea-vs_replace.txt
    GameExamples/SR3/8243432bd4dd0947-vs_replace.txt
    GameExamples/SR3/eca3e308dd737cb6-vs_replace.txt
    GameExamples/WatchDogs/73339ea29a5972eb-vs_replace.txt
    GameExamples/WatchDogs/e195cd9c9ca300fd-vs_replace.txt
    GameExamples/WatchDogs/fbe107d01dd27e4a-ps_replace.txt