Signature Parser: Parse globalFlags into Subshader Feature Info flags
Some shaders (notably shader model 4) with an SFI0 section are missing
the comment block and the features that fill this section out are
instead located in the globalFlags. However, globalFlags is also missing
some of the other features, so we can't just use one or the other.

This adds parsing of the globalFlags and combines it with the flags
obtained from the comment block to create the SFI0 section.

I've also done a little bit more testing and confirmed that the early
depth stencil flag that seemed to clash with the raw/structured buffer
flags does not show up in the SFI0 section, suggesting that the DX12
compiler's reference is more accurate.