Signature Parser: Force upgrade to SHEX and OSG1 when required
When certain features are in use, such as minimum precision or
raw/structured buffers we need to force the shader to use the extended
bytecode format, and if minimum precision is in use we need to force the
output signature section to a higher version.

All assembly test cases (except for the debug test cases) are now
passing again, and no longer require --lenient to do so.

The list of features that can force an upgrade may still be incomplete,
and it would be worthwhile writing test cases that focus purely on the
SFI flags to tease out any missing cases - but this should give us good
coverage regardless.