Work around vs2013 compiler bug causing crash at exit
The earlier attempt to drop the oddball precision debugging feature from
the assembler resulted in crashes at exit when built with the vs2013
toolchain, but not when built with vs2017.

Investigation showed that the class ParseError (as defined in
Assembler.cpp) destructor was being called from the standard library's
atexit handler, seemingly to destroy a statically allocated object -
however there was no statically allocated object associated with that
class, and as a result it clobbered nearby unrelated static memory
instead.

When the oddball debug feature was present it would clobber failFile
and nothing would notice since that was just a C style pointer that
nothing cared about. But when that was removed it would clobber codeBin
instead, which being a static C++ object had it's destructor scheduled
to be called from the same atexit handler shortly after the ParseError
destructor. At that point the memory corruption would become apparent
and we would get an actual crash.

However, this destructor should not have been called at all, because:

    class ParseError: public exception {
	    ...
    };
   ^^^^ There are no statically allocated objects here, nor anywhere
        else, nor was a ParseError constructed during execution.

However, SignatureParser.cpp also defines a class with the same type
name, and does statically allocate an object associated with it:

    class ParseError : public exception {} parseError;
                                           ^^^^^^^^^^ here.

So, the vs2013 compiler has managed to mix these two up, and is calling
the destructor for the variant in Assembler.cpp on a non-existent
statically allocated object due to an unrelated statically allocated
object in SignatureParser.cpp.

The solution is to give both classes separate names so that the compiler
cannot mix them up (potentially changing the namespace may also help,
but given that the .cpp files are separate compilation units and should
not be able to see each other's declarations this should not matter).

Notably this is not the first time we have come across similar issues in
the vs2013 compiler, where "SetShader" was called from the wrong class.