Assembler test suite: Add existing assembly only tests
Most of our assembly tests start as HLSL and get compiled, then a
disassembly + assembly validation is performed on the compiled shader.
A few of our tests start out as assembly to test instructions that we
are unable to produce with fxc, which are missing from the test suite.

For these ones we assemble the shader, disassemble the result and then
compare the two assembly listings. The comparison ignores comments and
spaces at the end of each line to account for minor differences between
disassemblers and to allow for RDEF and STAT to have been stripped
(although this means it won't currently check the signature comment
blocks)

Also renames the run_asm_test function to run_hlsl_asm_test to
distinguish between running an assembler test on hlsl vs assembly.