Assembler: Rework SPRG parsing to use a table of register names to file
This reworks the special purpose register parsing to use a table rather
than each having their use their own code. The table is sorted by "file"
number (not sure where the term "file" came from - maybe referring to a
register file index?) to make it easier to spot missing entries.

It sets the fields defined inside the encoded operand rather than
setting the entire operand at once, which allows for the extended flag
to be preserved eliminating the funny back up and restore strategy that
was used inconsistently in some of the registers.

It cleans up the swizzle handling and is more explicit - rather than "if
the bit at the start == entire operand" it now reads more "if there is a
swizzle" and gets rid of odd way of the setting the entire operand to
two different values based on this, now setting just the relevant field.

This fixes a minor issue in vGSInstanceID where the .x swizzle was
assumed (minor, as there is only a .x swizzle possible for that SPR, and
handleSwizzle would have corrected it if another had been used).

There is still some odd handling preserved from the original assembler
related to setting the correct value in comps_enum. We store the default
value to use for each SPR when there is no swizzle in the table, where
this is based on the subtle differences in how the assembler handled
each one before - 2 where the assembler unconditionally called
handleSwizzle (and where a bug would cause it to process the entire
operand as the swizzle if there was none, hence never hitting the code
path that sets comps_enum to 0), 1 where the assembler either didn't
call handleSwizzle at all, or where it did so conditionally, and 0 as a
special case for the "null" SPR.

I'm not certain that this is the best way to handle this - note that
the abovementioned bug combined with special handling in dcl_input
(that subtracts one from comps_enum when no swizzle is present,
effectively clearing it if it was set to 1, or changing 2s to 1s)
may interact with this, but this is at least passing all test cases.

This patch also makes some small tweaks the assembler test case script a
little to indicate which stage of the test failed, and all assembler
test cases are still passing after this change (except debug, which was
failing before).